LayoutTests/platform/qt/fast/xsl/xslt-recursion-expected.txt

CONSOLE MESSAGE: line 1: <html xmlns='http://www.w3.org/1999/xhtml/'><body><p>Running an XSL-T 1.0 stylesheet with a 2.0 processor.</p></body></html>
Success!
